RAMCC/club/requirements.txt
Django==2.1.3
django-crispy-forms==1.7.2
django-octicons==1.0.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
django-static-precompiler==1.8.2
djangorestframework==3.9.0
Pillow==5.3.0
pytz==2018.7
six==1.11.0
